Added function to retrieve general data related with current session
Also, all calls to urllib have been updated to make sure that this code works both in Python 2 and 3.

NOTE: I  have used try/except fallback system to use proper urllib classes depending on Python version we are using to run our code. There are some more elegant alternatives such as using future or six modules but this obligate us to include external packages and we want to minimzie that as much as possible. At least, for now.

https://python-future.org/compatible_idioms.html